Module battleship::app::hooks

Hooks to be added to the tide App.

Modules

app_status 🔒

board_status 🔒

drop_board 🔒

list_boards 🔒

list_strikes 🔒

List all the strikes that occurred on a board.

new_board 🔒

simulated_failure 🔒

Special module only enabled if feature simulate_failures is enabled.

strike 🔒

termination 🔒

Structs

AppStatusHook

[Endpoint] for creating a new board.

BoardStatusHook

[Endpoint] for creating a new board.

DropBoardHook

[Endpoint] for creating a new board.

ListBoardsHook

[Endpoint] for creating a new board.

ListStrikesHook

[Endpoint] for listing all the Strikes that occurred on a board.

NewBoardHook

[Endpoint] for creating a new board.

SimulatedFailureHook

Special Hook which is a [Endpoint], but only for the consumption of AppHook. This simulates a network failure should AppHook choose to do so.

StrikeHook

[Endpoint] for creating a new board.

TerminationHook

Functions

should_fail

Roll the dice to see if we should simulate a failure.