Module battleship::app::hooks::simulated_failure

Special module only enabled if feature simulate_failures is enabled.

This simulates network failures like Gateway Error or Request Timeouts, which would normally not happen on a localhost.

Structs

SimulatedFailureHook

Special Hook which is a [Endpoint], but only for the consumption of AppHook. This simulates a network failure should AppHook choose to do so.

Constants

REQUEST_TIMEOUT 🔒

Functions

should_fail

Roll the dice to see if we should simulate a failure.