Module battleship::config

Configuration module, containing constants, statics and HostConfiguration struct to store runtime configurations.

Modules

_serde_offset_date_time 🔒

host_config 🔒

serde_offset_date_time

Serialize and Deserialize module for DATETIME_FORMAT.

Structs

DATETIME_FORMAT

The default datetime format for use in this app.

HostConfiguration

A struct to parse the configurations from host.json, or whatever path config was set to in the command line arguments.

Constants

DEFAULT_ADDR

DEFAULT_BOARD_SIZE

DEFAULT_CONFIG_PATH

DEFAULT_PORT

DEFAULT_SHIP_COUNT

DEFAULT_SIMULATED_FAILURE_FACTOR

SHIP_GENERATION_MAX_RETRY

Maximum number of retries for each ship before we give up.