Enum battleship::error::AppError

pub enum AppError {

    LockPoisoned(&'static str),
    ConfigNotReadable {
        path: String,
        err: Error,
    },
    ConfigNotParsable {
        path: String,
        err: Error,
    },
    DebugOnlyFeature,
    ServerStartUpError {
        addr: String,
        err: Error,
    },
    TideError(Error),
    InvalidParameters(Error),
    MissingParameters(Vec<&'static str>),
    CtrlCError {
        message: String,
    },
    JSONSerializationError,
    DuplicatedBoard {
        uuid: Uuid,
    },
    MissingBoard {
        uuid: Uuid,
    },
    InvalidBoard {
        uuid_str: String,
    },
    FrozenBoard {
        uuid: Uuid,
    },
    RemoteRequestedTermination {
        message: String,
    },
    PositionOccupied {
        x: usize,
        y: usize,
    },
    CoordinatesOutOfBounds {
        x: usize,
        y: usize,
    },
    StrikeReportRequestedOnEmpty,
    Unknown {
        context: String,
    },

}

A collection of all possible errors raised by the app.

Variants

LockPoisoned(&'static str)

A lock for an object is poisoned; execution cannot continue.

This can occur when a thread panics while holding a writable guard reference to the lock.

This error is irrecoverable; please restart the server.

Status

500 INTERNAL SERVER ERROR

ConfigNotReadable

Fields

path: String

err: Error

host.json was not found at the specified location.

Status

500 INTERNAL SERVER ERROR

ConfigNotParsable

Fields

path: String

err: Error

host.json contains invalid JSON.

Status

500 INTERNAL SERVER ERROR

DebugOnlyFeature

The feature requested is only enabled in debug mode.

Status

406 NOT ACCEPTABLE

ServerStartUpError

Fields

addr: String

err: Error

The server failed to start. This is typically caused by the OS preventing the host from starting up due to port number clash etc.

Status

500 INTERNAL SERVER ERROR

TideError(Error)

[tide] encountered an error. The error message from [tide] will be displayed.

Status

Varies. The [tide::Error] status code is used.

InvalidParameters(Error)

A query parameter provided is invalid.

Status

400 BAD REQUEST

MissingParameters(Vec<&'static str>)

A mandatory query parameter is missing.

Status

404 NOT FOUND

CtrlCError

Fields

message: String

The OS prevented the host from listening to Ctrl-C termination commands.

Status

410 GONE

JSONSerializationError

A return value had been computed, but the object failed to serialize.

Status

500 INTERNAL SERVER ERROR

DuplicatedBoard

Fields

uuid: Uuid

The host is attempting to create a board with an existing Uuid.

This is an improbable cause of error; if you are unlucky enough to encounter this, simply retry the request.

Status

500 INTERNAL SERVER ERROR

MissingBoard

Fields

uuid: Uuid

The provided Uuid does not match any existing boards.

###ยStatus 404 NOT FOUND

InvalidBoard

Fields

uuid_str: String

The provided baord ID is not a well formatted Uuid.

###ยStatus 400 BAD REQUEST

FrozenBoard

Fields

uuid: Uuid

The target board cannot be modified, as the game has already ended.

Status

406 NOT ACCEPTABLE

RemoteRequestedTermination

Fields

message: String

A termination request had been sent with an error message. The host will now terminate with the message as its error.

Status

410 GONE

PositionOccupied

Fields

x: usize

y: usize

A requested position is not available.

Status

406 NOT ACCEPTABLE

CoordinatesOutOfBounds

Fields

x: usize

y: usize

A requested position is out of the bounds of the board. Check the board size before retrying.

Status

406 NOT ACCEPTABLE

StrikeReportRequestedOnEmpty

A strike report had been requested on an empty position that had not been struck before.

Internal error only; should be unreachable.

Status

500 INTERNAL SERVER ERROR

Unknown

Fields

context: String

An unknown error had occurred. Refer to the context for more information.

Status

500 INTERNAL SERVER ERROR

Implementations

impl AppError

pub fn error_code(&self) -> StatusCode

Get the HTTP Status code for this error.

Full list here: <https://docs.rs/tide/latest/tide/enum.StatusCode.html>

pub fn name(&self) -> String

Get the name of this error.