Module battleship::models

A collection of structs.

Modules

board 🔒

board_status 🔒

coordinates 🔒

orientation 🔒

Defines orientation on the board.

position 🔒

ship 🔒

ship_intel 🔒

ship_status 🔒

Defines orientation on the board.

ship_types 🔒

Possible types of ships.

strike 🔒

strike_report 🔒

traits

Structs

Board

BoardStatus

A brief summary of the status of the board.

Coordinates

Denotes the position of on a board.

Ship

ShipIntel

An abridged status of a Ship for the purpose of returning a global view of the ships on a Board.

Strike

Recording a Strike.

StrikeReport

A report for a strike, formatted for export.

Enums

Orientation

Orientation of the ship.

Position

Indicates the state of a position on the grid.

ShipStatus

Orientation of the ship.

ShipType

All possible ship types.

Constants

CHAR_EMPTY

CHAR_MISSED

CHAR_SHIP_HIT

CHAR_SHIP_UNHIT